Refactor vale components for improved structure and functionality
- Updated import paths for `ValeAlumno` and `ValeDetailsDisplay` components.
- Introduced new `AlumnoVale`, `ColumnsVale`, `DetailsVale`, `TableVale` components to enhance the vale management interface.
- Implemented a tabbed view in `AlumnoVale` for better state management of vales.
- Added detailed display functionality in `DetailsVale` for vale information.
- Created a table structure in `TableVale` to list vales with appropriate columns and filters.
