+++ src/components/layouts/vales/alumno/DetailsVale.tsx
