+++ src/components/layouts/vales/alumno/TableVale.tsx
@@ -3,7 +3,7 @@ import { useAlumnoVales } from "@/hooks/Vales/useVales";
 import { Table } from "@/components/table/Table";
 import { EstadoValeFilter } from "@/components/layouts/vales/alumno/FilterVale";
 import { EstadoVale, Vale } from "@/types/ValeTypes";
-import { columsAlumnoVale } from "./columns";
+import { columsAlumnoVale } from "./ColumnsVale";
 
 export const ValesAlumnoTable = ({ viewType }: { viewType: EstadoVale }) => {
   const { data, isLoading, isError } = useAlumnoVales(viewType);
